pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "modelcontextprotocol>=0.1.0",
    "numpy>=2.2.4",
    "openai>=1.74.0",
    "python-dotenv>=1.1.0",
]
